Moved StatePath type, renamed internal Sink -> Listener
